+++ src/simpletextview.h
@@ -8,16 +8,16 @@
 struct SimpleTextView;
 
 typedef struct {
-	void ( *SimpleTextView )( struct SimpleTextView*, char* text, u8, u8 );
+	void ( *SimpleTextView )( struct SimpleTextView*, char* text, s16, s16 );
 	void ( *destroy )( struct SimpleTextView* );
 	void ( *render )( struct SimpleTextView* );
 	void ( *position )( struct BaseView* );
 	void ( *renderChildren )( struct BaseView* );
 	void ( *addChildView )( struct BaseView*, struct BaseView* );
 	void ( *setPlane )( struct BaseView*, u16 );
-	void ( *placeTile )( struct BaseView*, u8, u8, u8, u16, bool, bool );
-	void ( *placeTileSeries )( struct BaseView*, u8, u8, u8, u8, u8, u16, bool );
-	bool ( *checkTileBoundary )( struct BaseView*, u8, u8 );
+	void ( *placeTile )( struct BaseView*, s16, s16, u8, u16, bool, bool );
+	void ( *placeTileSeries )( struct BaseView*, s16, s16, s16, s16, u8, u16, bool );
+	bool ( *checkTileBoundary )( struct BaseView*, s16, s16 );
 } SimpleTextView_vtable;
 
 typedef struct SimpleTextView {
@@ -26,7 +26,7 @@ typedef struct SimpleTextView {
 	char* text;
 } SimpleTextView;
 
-void SimpleTextView_ctor( SimpleTextView* this, char* text, u8 x, u8 y );
+void SimpleTextView_ctor( SimpleTextView* this, char* text, s16 x, s16 y );
 void SimpleTextView_dtor( SimpleTextView* this );
 void SimpleTextView_render( SimpleTextView* this );
